+++ CHANGES.md
@@ -11,6 +11,16 @@
 
 ## develop
 
+- [ADD] `StblBox` の情報へのアクセスを簡単かつ安全にするための `SampleTableAccessor` 構造体を追加する
+  - @sile
+- [ADD] `SttsBox::from_sample_deltas()` 関数を追加する
+  - @sile
+- [ADD] `Utf8String::into_null_terminated_bytes()` メソッドを追加する
+  - @sile
+- [CHANGE] 仕様上 0 を取らないフィールドの型は `NonZeroXXX` にする
+  -  @sile
+- [UPDATE] ボックスに `Hash` を実装する
+  - @sile
 - [CHANGE] `BoxHeader` 書き込み時に large size にするかどうかの自動判定は行わないようにする
   - `BoxSize` 自体はどちらを使うべきかの情報を有しているので、それをそのまま反映するようにした
   - @sile